Changed the structured and fixed glitches
