+++ OverloadingFixed/__init__.py
@@ -1,0 +1,19 @@
+from .main import *
+"""
+--------------
+overloading.py
+--------------
+
+Function overloading for Python 3
+
+* Project repository: https://github.com/toto112358/overloading.py
+* Documentation: https://overloading.readthedocs.org/
+
+Copyright © 2014–2016 Kalle Tuure. Released under the MIT License.
+Copyright © 2021–2021 L. Pham-Trong Released under the MIT License.
+
+"""
+
+__version__ = '1.0'
+
+__all__ = ['overload', 'overloaded', 'overloads']
